requirements.txt
pytest
requests
jsonschema
allure-pytest
selene==2.0.0rc9